requirements.txt
Pillow
urllib3
python-dotenv